change the project structure to get the new  database strcture change the update to create

+++ src/main/java/com/app/sorteios/api/repository/ClubRepository.java
@@ -4,7 +4,7 @@ import com.app.sorteios.api.model.Club;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import org.springframework.data.jpa.repository.JpaRepository;
 @Repository
-public interface ClubRepository extends JpaRepository<Club, Integer> {
-    // You can define custom queries or methods here if needed
+public interface ClubRepository extends JpaRepository<Club, Long> {
 }
