+++ src/main/java/com/app/sorteios/api/repository/CoachRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CoachRepository extends JpaRepository<Coach, Integer> {
-    // You can define custom queries or methods here if needed
+public interface CoachRepository extends JpaRepository<Coach, Long> {
 }
+
