+++ src/main/java/com/app/sorteios/api/repository/RefereeRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RefereeRepository extends JpaRepository<Referee, Integer> {
+public interface RefereeRepository extends JpaRepository<Referee, Long> {
 }
+
